Запущены тесты

+++ main-project/catalog.h
@@ -1,13 +1,13 @@
 #pragma once
 #ifndef CATALOG_H
 #define CATALOG_H
-
 #include "constants.h"
+#include <string>
 
 
 struct catalog
 {
-    unsigned double cost;
+    double cost;
     unsigned int quan;
     char category[MAX_STRING_SIZE];
     char name[MAX_STRING_SIZE];
